applications/backend/requirements.txt
fastapi
uvicorn
azure-cosmos
python-dotenv
